Move AssignmentCreationViewModel out of AssignmentCreationFragment
